Add toolbar button icon for logging action
The upstream code had  stock=gtk.STOCK_JUMP_TO for the icon
However, if we use the stock attribute - we get the following error
TypeError: action() got an unexpected keyword argument 'stock'

So that seems to now be replaced with the 'icon' attribute.
But we still get the following error:
  @action(_('Log Entry'), icon=gtk.STOCK_JUMP_TO,
  menuhints='view',accelerator='<Primary><Shift>J') # T: menu item

    File "gi/__init__.py", line 69, in __getattr__
    AttributeError: When using gi.repository you must not import static
    modules like "gobject".
     Please change all occurrences of "import gobject" to "from
     gi.repository import GObject".

Got a hint about that error from:
https://github.com/rabbitvcs/rabbitvcs/pull/189/commits/1ced574155ed3068de9d67adff6424ada9984d34

Replaced import gtk
with
from gi.repository import Gtk as gtk

And finally, we get a Log Entry icon in the toolbar!